В сущностях поменял BigInteger, добавил аннотацию ToString. Добавил в мапперы для сущностей получение дто из сущности. В репозиториях и сервисах поменял BigInteger на Long, добавил методы. Создал слои validator, utils, handler, exception, config и aop. Создал 4 контроллера

+++ transfer/src/main/java/com/bank/transfer/dto/AccountTransferDto.java
@@ -1,14 +1,10 @@
 package com.bank.transfer.dto;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
-/**
- * Класс, описывающий Data Transfer Objects Account Transfer
- * @author Savenkov Artem
- */
-public record AccountTransferDto(BigInteger id,
-                                 BigInteger accountNumber,
-                                 BigInteger amount,
+public record AccountTransferDto(Long id,
+                                 Long accountNumber,
+                                 BigDecimal amount,
                                  String purpose,
-                                 BigInteger accountDetailsId) {
+                                 Long accountDetailsId) {
 }
